人工智能基础/实验/lab2/LAB2_PB18111791_雷雨轩/requirement.txt
python==3.6
pytorch=1.8.1
torchvision=0.9.1
cvxopt==1.2.6
matplotlib==3.3.4

